+++ src/main/resources/webapp/docx-exporter/html/popupForm.html
@@ -1,5 +1,5 @@
 <div class="form-wrapper">
-    <div class="docx-in-progress-overlay"><img src='/polarion/ria/images/progressWheel48.svg' alt=""/><span id="docx-in-progress-message"></span></div>
+    <div class="docx-in-progress-overlay"><span class="sbb-spinner" role="img" aria-label="Loading" style="width:48px;height:48px;"></span><span id="docx-in-progress-message"></span></div>
 
     <div class="docx-notifications">
         <div class="alert alert-warning" style="display: none"></div>
