+++ src/main/resources/webapp/docx-exporter/html/sidePanelContent.html
@@ -174,7 +174,7 @@
         <button type='button' id='export-docx'>
             <img class='append-build-number' src='/polarion/ria/images/dle/operations/actionMsWordRoundtrip16.svg' alt=''/>Export to DOCX
         </button>
-        <img id='export-docx-progress' src='/polarion/ria/images/progressWheel48.svg' alt=''/>
+        <span id='export-docx-progress' class='sbb-spinner' role='img' aria-label='Loading'></span>
         <div id='export-docx-error'></div>
         <div id='export-docx-warning'></div>
     </div>
